+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog for NeoFS Node
 - Initial support for meta-on-chain for objects (#2877)
 
 ### Fixed
+- `neofs-cli object delete` command output (#3056)
 
 ### Changed