+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changelog for NeoFS Node
 ### Fixed
 
 ### Changed
+- Number of cuncurrenly handled notifications from the chain was increased from 10 to 300 for IR (#3068)
 
 ### Removed